+++ src/Component/Contract/ClientInterface.php
@@ -3,7 +3,7 @@
 namespace Heptacom\AdminOpenAuth\Component\Contract;
 
 use Heptacom\AdminOpenAuth\Struct\TokenPairStruct;
-use Heptacom\AdminOpenAuth\Struct\UserStruct;
+use Heptacom\OpenAuth\Struct\UserStruct;
 use League\OAuth2\Client\Provider\AbstractProvider;
 
 interface ClientInterface