+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Move UserStruct into Heptacom\OpenAuth\Struct namespace and removed Shopware dependency
+
 # 2.0.0
 
 * Fix typo in administration template registration